+++ graph/db/test_postgres.go
@@ -6,14 +6,12 @@ import (
 	"database/sql"
 	"testing"
 
-	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/lightningnetwork/lnd/sqldb"
-	"github.com/stretchr/testify/require"
 )
 
-// NewTestDB is a helper function that creates a SQLStore backed by a postgres
-// database for testing.
-func NewTestDB(t testing.TB) V1Store {
+// newBatchQuerier creates a new BatchedSQLQueries instance for testing
+// using a PostgreSQL database fixture.
+func newBatchQuerier(t testing.TB) BatchedSQLQueries {
 	pgFixture := sqldb.NewTestPgFixture(
 		t, sqldb.DefaultPostgresFixtureLifetime,
 	)
@@ -23,18 +21,9 @@ func NewTestDB(t testing.TB) V1Store {
 
 	db := sqldb.NewTestPostgresDB(t, pgFixture).BaseDB
 
-	executor := sqldb.NewTransactionExecutor(
+	return sqldb.NewTransactionExecutor(
 		db, func(tx *sql.Tx) SQLQueries {
 			return db.WithTx(tx)
 		},
 	)
-
-	store, err := NewSQLStore(
-		&SQLStoreConfig{
-			ChainHash: *chaincfg.MainNetParams.GenesisHash,
-		}, executor,
-	)
-	require.NoError(t, err)
-
-	return store
 }
